Core Cryptosystem

Paillier.Paillier — Module

module Paillier

CurrentModule = Paillier

DocTestSetup = quote
    using Paillier
    publickey, privatekey = generate_paillier_keypair(256)
end

Basic Usage

Generate a public and private key of given length:

using Paillier
publickey, privatekey = generate_paillier_keypair(1024);

The publickey is used for encryption via encrypt, and privatekey is required for decryption decrypt.

Encrypted numbers can be added together:

julia> a = encrypt(publickey, 10);

julia> b = encrypt(publickey, 20);

julia> decrypt(privatekey, a + b)
30

Methods

Paillier.generate_paillier_keypair — Function

generate_paillier_keypair([rng=GLOBAL_RNG], n_length=2048)

Generate a new Paillier keypair of given bit length.

Returns a 2-tuple of the public and private key.

Paillier.encrypt — Function

encrypt(public_key, message)

Encrypt a message with a given public key and return an Encrypted. The message must be a positive integer under public_key.n - following Julia's Int64 type larger numbers will wrap around and not throw an error.

See decrypt if you'd like your original message back.

encrypt(::Array{Encoded}, ::Encoding)
encrypt(::Array{Encoded}, ::PublicKey)

Encrypt an array of encoded instances

Paillier.decrypt — Function

decrypt(private_key, ciphertext::Ciphertext)
decrypt(private_key, ciphertext::Encrypted)

The inverse of encrypt, uses the private_key to decrypt an encrypted message (ciphertext) to a positive integer less than public_key.n. The result is always a BigInt.

Examples

DocTestSetup = quote
    using Paillier
    publickey, privatekey = generate_paillier_keypair(256)
end

julia> ciphertext = encrypt(publickey, 10);

julia> decrypt(privatekey, ciphertext)
10

Paillier.encrypt_raw — Function

encrypt_raw(public_key, message)

Internal version of encrypt that returns the raw Ciphertext

which is just a BigInt. Note this includes obfuscation so

a directly encrypted unsigned integer is safe to share.

Paillier.obfuscate — Function

obfuscate(encrypted)
obfuscate(rng, encrypted)
obfuscate(pub, ciphertext)
obfuscate(rng, pub, ciphertext)

Salt the Encrypted with a new random number. Required before sharing ciphertexts with another party.

Calling with a Ciphertext will return a new obfuscated Ciphertext, and calling with an [Encrypted])@ref will return a new Encrypted.

Paillier.decrease_exponent_to — Function

decrease_exponent_to(enc, new_exponent)

Compute an EncryptedNumber with the same value but a lower exponent.

We multiply the encoded value by the Encoding.base and decrement the EncryptedNumber.exponent - such that the decoded value remains the same.

We can keep ratcheting down the EncryptedNumber's exponent until the encoded integer overflows. This overflow may not be caught.

There is also a (much faster) version that acts on an Encodeds.

Homomorphic Operations

Base.:+ — Method

+(::Encrypted, ::Number)
+(::Encrypted, ::Encrypted)

Homomorphic addition. Add an Encrypted to either a plaintext or another encrypted.

Note a plaintext number will be automatically encrypted and therefore must meet the same conditions outlined in encrypt - i.e. positive integers under public_key.n.

Example

DocTestSetup = quote
    using Paillier
    publickey, privatekey = generate_paillier_keypair(256)
end

c1 = encrypt(publickey, 10)
c2 = encrypt(publickey, 1000);
println(decrypt(privatekey, c1 + 90))
println(decrypt(privatekey, c1 + c2))

# output

100
1010

Base.:* — Method

*(::Encrypted, ::Integer)

Homomorphic multiplication. Allows the multiplication of an Encrypted and cleartext integer.

Example

DocTestSetup = quote
    using Paillier
    publickey, privatekey = generate_paillier_keypair(256)
end

a = encrypt(publickey, 10)
println(decrypt(privatekey, a * 10))

# output

100

Types

Paillier.PublicKey — Type

PublicKey(p, q)
PublicKey(n)

A Paillier cryptosystem public key. Create a keypair with generate_paillier_keypair

Paillier.PrivateKey — Type

PrivateKey(public_key, p, q)
PrivateKey(public_key, p, q)

A Paillier cryptosystem private key. Create a keypair with generate_paillier_keypair.

Paillier.Encrypted — Type

Encrypted(ciphertext, public_key)
Encrypted(ciphertext, public_key, is_obfuscated::Bool)

An Encrypted is the Paillier.jl library's low level encrypted type. This object that includes the ciphertext, public_key and tracks whether obfuscation has occurred (assumed as false if not provided).

Paillier.Ciphertext — Type

Ciphertext

The raw encrypted information is always a Ciphertext which is simply an alias of BigInt.