Encoding Floats

Encoding Floats

Types

Paillier.Encoding — Type.

Encoding(::DataType, ::PublicKey)
Encoding(::DataType, ::PublicKey, base::Int64)

A datatype for describing an encoding scheme.

The public key is included as the encoding is effected by the maximum representable integer which varies with the public_key. Although I could be convinced to change this.

Examples

Setting a base value is optional:

julia> encoding = Encoding(Float64, public_key, 64)

Full example showing homomorphic operations on floating point numbers:

julia> keysize = 2048
julia> publickey, privatekey = generate_paillier_keypair(keysize)
julia> encoding = Encoding(Float32, publickey)
julia> a = Float32(π)
julia> enc1 = encode_and_encrypt(a, encoding)
julia> decrypt_and_decode(privatekey, enc1)
3.1415927f0
julia> enc1.exponent
-6
julia> b = 100
julia> enc2 = encode_and_encrypt(b, encoding)
julia> decrypt_and_decode(privatekey, enc1 + enc2)
103.141594f0
julia> decrypt_and_decode(privatekey, enc1 - 20.0)
-16.858408f0

source