Core Cryptosystem

Types

Paillier.Encrypted — Type

Encrypted(ciphertext, public_key)
Encrypted(ciphertext, public_key, is_obfuscated::Bool)

An Encrypted is the Paillier.jl library's low level encrypted type. This simple object that includes the ciphertext, public_key and tracks whether obfuscation has occurred (assumed as false if not provided).

Methods

Paillier.encrypt — Function

encrypt(public_key, message)

Encrypt a message with a given public key and return an Encrypted. The message must be a positive integer under public_key.n - following Julia's Int64 type larger numbers will wrap around and not throw an error.

See decrypt if you'd like your original message back.

encrypt(::Array{Encoded}, ::Encoding)
encrypt(::Array{Encoded}, ::PublicKey)

Encrypt an array of encoded instances

Paillier.decrypt — Function

decrypt(private_key, ciphertext::Ciphertext)
decrypt(private_key, ciphertext::Encrypted)

The inverse of encrypt, uses the private_key to decrypt an encrypted message (ciphertext) to a positive integer less than public_key.n. The result is always a BigInt.

Examples

julia> publickey, privatekey = generate_paillier_keypair(128);
julia> ciphertext = encrypt(publickey, 10);
julia> decrypt(privatekey, ciphertext)
10

Paillier.encrypt_raw — Function

encrypt_raw(public_key, message)

Internal version of encrypt that returns the raw Ciphertext - which is just a BigInt. Note this includes obfuscation so a directly encrypted unsigned integer is safe to share.

Paillier.obfuscate — Function

obfuscate(encrypted)
obfuscate(rng, encrypted)

Salt the Encrypted with a new random number. Required before sharing ciphertexts with another party.