Encoding

Paillier.jl allows encoding of Julia primitive numbers. The following example shows carrying out homomorphic operations on floating point numbers - in this case Float32.

julia> encoding = Encoding{Float32}(publickey);

julia> a = Float32(π)
3.1415927f0

julia> enc1 = encode_and_encrypt(a, encoding);

julia> decrypt_and_decode(privatekey, enc1)
3.1415927f0

julia> enc1.exponent
-6

julia> enc2 = encode_and_encrypt(100, encoding);

julia> decrypt_and_decode(privatekey, enc1 + enc2)
103.141594f0

julia> decrypt_and_decode(privatekey, enc1 - 20.0)
-16.858408f0

Note the enc1.exponent is a public number which reveals size information about the encrypted value.

API

Paillier.encode_and_encrypt — Function

encode_and_encrypt(plaintext::Number, encoding::Encoding)
encode_and_encrypt(plaintext::Number, encoding::Encoding, exponent::Int64)

Encode the plaintext number using given encoding and encrypt using the PublicKey from the encoding.

encode_and_encrypt(xs::Array{<:Number}, encoding::Encoding)
encode_and_encrypt(xs::Array{<:Number}, encoding::Encoding, exponent::Int64)

Create an EncryptedArray of your plaintext numbers.

Paillier.decrypt_and_decode — Function

decrypt_and_decode(privatekey::PrivateKey, enc::EncryptedNumber)

Decrypt an EncryptedNumber using the given PrivateKey and decode it using the EncryptedNumber's encoding.

Paillier.max_int — Function

max_int(::PublicKey)

The maximum signed integer for our encoding system based on the public key.

We use a full third of the range for overflow detection.

Paillier.encode — Function

encode(::Number, ::Encoding)
encode(::Number, ::Encoding, exponent::Int64)

Encode a number but don't encrypt it for the given Encoding - producing an Encoded. See encode_and_encrypt for a method that also encrypts. If the exponent is not provided we attempt to match the precision of the passed julia type. See decode to go the other direction.

Paillier.decode — Function

decode(encoded::Encoded)
decode(encoded::BigInt, exponent::Int64, encoding::Encoding)

The inverse of encode, computes the decoding of the Encoded integer representation.

Types

Paillier.Encoding — Type

Encoding{::DataType}(::PublicKey)
Encoding{::DataType,(::PublicKey, base::Int64)

A datatype for describing a fixed point encoding scheme for Julia DataTypes.

The public key is required as the maximum representable integer for an Encoding varies with the public_key.

Setting a base value is optional - other Paillier implementations may use a different base.

To work with negative and floating point numbers we follow the encoding scheme of python-paillier.

Example

An encoding for Float64 numbers specifying the optional base for encoding a Float64:

julia> encoding = Encoding{Float64}(public_key, 64)

Paillier.Encoded — Type

A datatype for a plaintext encoded number. Returned by the raw encode methods, most users should instead use encode_and_encrypt.

Represents the Julia value as a BigInt.

Paillier.EncodedArray — Type

EncodedArray

A vector version of Encoded

Paillier.EncryptedNumber — Type

EncryptedNumber(::Encrypted, ::Encoding, exponent::Int64)
EncryptedNumber(::Encoded, ::PublicKey)

Datatype for representing an Encrypted number with a known Encoding.

DocTestSetup = quote
    using Paillier
    publickey, privatekey = generate_paillier_keypair(256)
    encoding = Encoding{Float64}(publickey, 64)
end

Example

encoded_number = encode(23.4, encoding)
encrypted_number = EncryptedNumber(encoded_number, publickey)
decrypt_and_decode(privatekey, encrypted_number)

# output

23.4

User Defined Encoding

Say you wanted to carry out Partially Homomorphic operations on values with uncertainty using the fantastic Meaurements package.

Do achieve this you need to define your Encoding type, and add methods to encode and decode the new Encoding type.

import Paillier
using Measurements

keysize = 2048
base = 64

MyType = Measurement{Float16}
publickey, privatekey = Paillier.generate_paillier_keypair(keysize)
encoding = Paillier.Encoding{MyType}(publickey, base)

# Support encoding any Measurement by encoding the value and error separately
# This will create an EncodedArray
function Paillier.encode(scalar::MyType, encoding::Paillier.Encoding{Measurement{T}}) where T
    internal_encoding = Paillier.Encoding{T}(encoding.public_key, encoding.base)
    encodings = Paillier.encode([scalar.val, scalar.err], internal_encoding)
    # create a copy of the EncodedArray with our Measurement encoding type
    Paillier.EncodedArray(encodings.plaintexts, encoding)
end

function Paillier.decode(encoded::Paillier.EncodedArray, exponent::Int64, encoding::Paillier.Encoding{Measurement{T}}) where T
    internal_encoding = Paillier.Encoding{T}(encoding.public_key, encoding.base)
    return measurement(
        Paillier.decode(encoded.plaintexts[1].value, exponent, internal_encoding),
        Paillier.decode(encoded.plaintexts[2].value, exponent, internal_encoding)
    )
end

"""
Because we are end up with an array of encrypted numbers
we may want to override some of the broadcast/array functionality
"""
+(enc_a::Paillier.EncryptedArray, plaintext::Measurement) = enc_a + Paillier.encode(plaintext, enc_a.encoding)

a = Measurement{Float16}(2000 ± 10)
b = Measurement{Float16}(100 ± 1)

enc1 = Paillier.encode_and_encrypt(a, encoding)
enc2 = Paillier.encode_and_encrypt(b, encoding)

enc3 = enc1 + enc2

c = Paillier.decrypt_and_decode(privatekey, enc3)
println("Adding encrypted Measurement numbers: D(E($a) + E($b)) = $c")

# Dircetly use our previously defined encoding function
encoded_b = Paillier.encode(b, enc3.encoding)
d = Paillier.decrypt_and_decode(privatekey, enc3 + encoded_b)
println("Adding encrypted number with encoded but unencrypted number (with uncertainty): D(E($c) + $b) = $d")

# Dircetly add a non encoded Measurement number
d = Paillier.decrypt_and_decode(privatekey, enc3 + b)
println("Adding encrypted number with Meaurement number: D(E$c) + $b) = $d")

# Subtraction
enc4 = Paillier.decrypt_and_decode(privatekey, enc3 - Paillier.encode_and_encrypt(a, enc3.encoding))
println("Subtract a constant (with uncertainty) from an encrypted number: D(E($c) - $a) = $enc4")

# Multiplication
enc5 = Paillier.decrypt_and_decode(privatekey, 3*enc1)
println("Scaling an encrypted Measurement number: 3 * $a = $enc5")